+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Cataloguing: identify video location from Apple QuickTime metadata, and 3GPP `loci` atom
 - Collection: stack RAW and HEIC with same file names
 - display home tile in side drawer when customized
+- Galician translation (thanks rcasl)
 
 ### Changed